+++ src/Authentication/Authentication.Core/Microsoft.Graph.Authentication.Core.csproj
@@ -16,10 +16,10 @@
     <PackageReference Include="Azure.Identity" Version="1.7.0" />
     <PackageReference Include="Azure.Identity.BrokeredAuthentication" Version="1.0.0-beta.3" />
     <PackageReference Include="Microsoft.Graph.Core" Version="2.0.14" />
-    <PackageReference Include="Microsoft.Identity.Client" Version="4.54.1" />
-    <PackageReference Include="Microsoft.Identity.Client.Broker" Version="4.54.1" />
+    <PackageReference Include="Microsoft.Identity.Client" Version="4.52.0" />
+    <PackageReference Include="Microsoft.Identity.Client.Broker" Version="4.52.0" />
     <PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="2.24.0" />
-    <PackageReference Include="Microsoft.Identity.Client.NativeInterop" Version="0.13.8" />
+    <PackageReference Include="Microsoft.Identity.Client.NativeInterop" Version="0.13.7" />
     <!--Always ensure this version matches the versions of System.Security.Cryptography.* dependencies of Microsoft.Identity.Client when targeting PowerShell 6 and below.-->
     <PackageReference Include="System.Security.Cryptography.ProtectedData" Version="6.0.0" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
